backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2

# LLM providers
anthropic==0.34.2
openai==1.47.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
